Updated preprocessing pipeline after second meeting
The data pre-processing pipeline was updated according to the last meeting. A detailed description on the updates is provided in RelatedToTransport.md. The main changes are the addition of more datasets related to transportation, making the preprocessing general so they can be applied on all datasets with a single pipeline, and joining all datasets by country code and year.